requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
fastapi==0.75.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylama==8.3.8
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.20.0
requests==2.27.1
sniffio==1.2.0
snowballstemmer==2.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
